+++ components/posts-page-component.js
@@ -1,9 +1,9 @@
 import { USER_POSTS_PAGE } from "../routes.js";
 import { renderHeaderComponent } from "./header-component.js";
-import { posts, goToPage } from "../main.js";
-import { likeEventListener } from "./like-component.js";
+import { posts, goToPage, getToken, renderApp, setPosts } from "../main.js";
 import { formatDistanceToNow } from "date-fns";
 import { ru } from "date-fns/locale";
+import { setLike, removeLike, getPosts } from "../api.js";
 
 export function renderPostsPageComponent({ appEl }) {
 
@@ -68,6 +68,47 @@ export function renderPostsPageComponent({ appEl }) {
       });
     });
   }
+
+  const likeEventListener = () => {
+    const likeButtons = document.querySelectorAll(".like-button");
+
+    likeButtons.forEach(likeButton => {
+      likeButton.addEventListener("click", (event) => {
+        event.stopPropagation();
+        const postId = likeButton.dataset.postId;
+        const index = likeButton.dataset.index;
+        likeButton.classList.add("shake-bottom");
+
+        if (posts[index].isLiked) {
+          removeLike({ token: getToken(), postId })
+            .then(() => {
+              posts[index].isLiked = false;
+            })
+            .then(() => {
+              getPosts({ token: getToken() })
+                .then((response) => {
+                  setPosts(response);
+                  likeButton.classList.remove("shake-bottom");
+                  renderApp();
+                })
+            })
+        } else {
+          setLike({ token: getToken(), postId })
+            .then(() => {
+              posts[index].isLiked = true;
+            })
+            .then(() => {
+              getPosts({ token: getToken() })
+                .then((response) => {
+                  setPosts(response);
+                  likeButton.classList.remove("shake-bottom");
+                  renderApp();
+                })
+            })
+        }
+      });
+    });
+  };
 
   likeEventListener();
 }
